requirements.txt
Django==5.0.6
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
djangorestframework-camel-case==1.0.2
django-cors-headers==3.14.0
